+++ src/Page/Crud/FormPage.php
@@ -33,8 +33,8 @@ use MB\Bitrix\AdminKit\Relation\EntityObjectFormSaver;
 use MB\Bitrix\AdminKit\Security\PermissionContext;
 use MB\Bitrix\AdminKit\Support\AdminKitJs;
 use MB\Bitrix\AdminKit\Support\DataWrapper;
-use MB\Bitrix\AdminKit\Support\ExceptionDiagnostics;
 use MB\Bitrix\AdminKit\Support\Enums\PageType;
+use MB\Bitrix\AdminKit\Support\ExceptionDiagnostics;
 use MB\Bitrix\AdminKit\Support\ResponseTerminator;
 use Throwable;